feat(pages): create Nordic landing page hero
- Built new landing page at app/page.tsx with Nordic design principles
- Moved profile management to dedicated /profile route
- Applied Nordic spacing scale throughout (spacing-6, spacing-20, etc.)
- Used Nordic color palette (fjord, forest, clay) for features and CTAs
- Implemented Nordic typography scale (heading-xl, heading-lg, heading-sm, body, small)
- Added soft gradient background with subtle Nordic color blurs
- Created responsive hero section with generous whitespace
- Built 3-column feature grid highlighting core capabilities
- Added bottom CTA section with fjord-600 background
- Updated navigation links to point to /profile instead of /

## Design Highlights
- Generous padding and spacing for breathing room
- Soft border radius on feature icons
- Subtle background gradients using Nordic colors
- Typography hierarchy using CSS custom properties
- Fully responsive with mobile-first approach
- Accessible color contrast (WCAG AA compliant)

Closes #14
